Move navigation components to organized directories
- Created and populated navigation directory with menu, menubar, and tabs components
- Added proper README documentation for each component
- Created consistent barrel exports (index.ts files)
- Added deprecation notices to original component locations
- Updated COMPONENT-MOVEMENT-PLAN.md with progress tracking
