+++ src/components/ui/navigation/index.ts
@@ -1,0 +1,3 @@
+export * from './menu';
+export * from './menubar';
+export * from './tabs';
