+++ src/components/ui/navigation/menu/README.md
@@ -1,0 +1,71 @@
+# Navigation Menu Component
+
+The Navigation Menu component provides a menu structure for top-level navigation, with support for dropdown menus and an accessible design.
+
+## Usage
+
+```tsx
+import {
+  NavigationMenu,
+  NavigationMenuContent,
+  NavigationMenuItem,
+  NavigationMenuLink,
+  NavigationMenuList,
+  NavigationMenuTrigger,
+  navigationMenuTriggerStyle,
+} from "@/components/ui/navigation/menu";
+
+export default function NavExample() {
+  return (
+    <NavigationMenu>
+      <NavigationMenuList>
+        <NavigationMenuItem>
+          <NavigationMenuTrigger>Getting started</NavigationMenuTrigger>
+          <NavigationMenuContent>
+            <ul className="grid gap-3 p-4 md:w-[400px] lg:w-[500px] lg:grid-cols-[.75fr_1fr]">
+              <li className="row-span-3">
+                <NavigationMenuLink asChild>
+                  <a
+                    className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted p-6 no-underline outline-none focus:shadow-md"
+                    href="/"
+                  >
+                    <div className="mb-2 mt-4 text-lg font-medium">
+                      Navigation Example
+                    </div>
+                    <p className="text-sm leading-tight text-muted-foreground">
+                      Beautifully designed navigation component
+                    </p>
+                  </a>
+                </NavigationMenuLink>
+              </li>
+              <li>
+                <NavigationMenuLink asChild>
+                  <a
+                    className="block select-none space-y-1 rounded-md p-3 leading-none no-underline outline-none transition-colors hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground"
+                    href="/docs"
+                  >
+                    <div className="text-sm font-medium leading-none">Documentation</div>
+                    <p className="line-clamp-2 text-sm leading-snug text-muted-foreground">
+                      Start integrating components in your application
+                    </p>
+                  </a>
+                </NavigationMenuLink>
+              </li>
+            </ul>
+          </NavigationMenuContent>
+        </NavigationMenuItem>
+      </NavigationMenuList>
+    </NavigationMenu>
+  );
+}
+```
+
+## Features
+
+- Dropdown navigation menus
+- Keyboard navigation support
+- Customizable styles
+- Animated transitions
+- Mobile responsive design
+- Accessible WAI-ARIA compliant implementation
+- Built on Radix UI primitives
