+++ src/components/ui/navigation/menu/index.ts
@@ -1,0 +1,1 @@
+export * from './navigation-menu';
