+++ src/components/ui/navigation/menubar/index.ts
@@ -1,0 +1,1 @@
+export * from './menubar';
