+++ src/components/ui/navigation/tabs/index.ts
@@ -1,0 +1,1 @@
+export * from './tabs';
